Refactor cherami-admin and cherami-cli code
